Split oversized sync-env source and rotation test files
Extract two cohesive modules out of src/sync-env.ts (586 lines, over the
400 cap):
- src/lib/sync-env-args.ts — Options type, USAGE text, and parseArgs.
- src/lib/sync-env-init.ts — findDevSource, validateInitConfig, and
  resolveAutoInit (the --init detection/validation logic).

sync-env.ts re-exports parseArgs and Options so the public surface the
tests import is unchanged; it drops to 299 lines.

Split src/__tests__/sync-env-tests/run-rotate.test.ts (1050 lines, over the
600 cap) along its existing section boundaries into three files, each well
under the cap:
- run-rotate.test.ts — core --rotate-keys forwarding (502 lines).
- run-rotate-init-auto.test.ts — --init detection/forwarding (447 lines).
- run-rotate-refresh.test.ts — --refresh-previews (159 lines).

Pure structural refactor: no behavior change. The full 83-test suite and
tsc --noEmit both pass unchanged.
